uuid: move into libcore
libuuid used to be a separate library since
de11d68a48e397b555731be04ba695b2ba1a0ccf ("CMake: refactor
dependencies of internal libraries").

Unclear what was it done for. The commit says "fir unit tests".
But they perfectly fine can depend on libcore like many of them
do already.

Because of libuuid being a separate library, libcore can't use
tt_uuid, because that would be a cyclic dependency. And that won't
allow to introduce tt_uuid-dependent API in a next patch.

Hence libuuid is merged into libcore.

Needed for #5568
