+++ src/CMakeLists.txt
@@ -190,7 +190,7 @@ endif()
 set_source_files_compile_flags(${server_sources})
 add_library(server STATIC ${server_sources})
 add_dependencies(server build_bundled_libs)
-target_link_libraries(server core coll http_parser bit uri uuid swim swim_udp
+target_link_libraries(server core coll http_parser bit uri swim swim_udp
                       swim_ev crypto mpstream crc32 tzcode)
 
 # Rule of thumb: if exporting a symbol from a static library, list the
@@ -220,7 +220,6 @@ if (TARGET_OS_FREEBSD AND NOT TARGET_OS_DEBIAN_FREEBSD)
     endif()
 endif()
 
-set (common_libraries ${common_libraries} ${LIBUUID_LIBRARIES})
 set (common_libraries ${common_libraries} PARENT_SCOPE)
 
 add_subdirectory(lib)
