+++ src/box/tuple.h
@@ -34,8 +34,8 @@
 #include "say.h"
 #include "diag.h"
 #include "error.h"
-#include "uuid/tt_uuid.h" /* tuple_field_uuid */
 #include "tt_static.h"
+#include "tt_uuid.h"
 #include "tuple_format.h"
 
 #if defined(__cplusplus)
