+++ src/lib/CMakeLists.txt
@@ -8,7 +8,6 @@ add_subdirectory(json)
 add_subdirectory(uri)
 add_subdirectory(http_parser)
 add_subdirectory(core)
-add_subdirectory(uuid)
 add_subdirectory(coll)
 add_subdirectory(crypto)
 add_subdirectory(swim)
