+++ src/lib/core/tt_uuid.h
