+++ src/lib/mpstream/CMakeLists.txt
@@ -1,2 +1,2 @@
 add_library(mpstream STATIC mpstream.c)
-target_link_libraries(mpstream core uuid ${MSGPUCK_LIBRARIES})
+target_link_libraries(mpstream core ${MSGPUCK_LIBRARIES})
