+++ src/lua/msgpack.c
@@ -42,10 +42,10 @@
 #include <small/region.h>
 #include <small/ibuf.h>
 
+#include "decimal.h" /* decimal_unpack() */
 #include "lua/decimal.h" /* lua_pushdecimal() */
-#include "lib/core/decimal.h" /* decimal_unpack() */
-#include "lib/uuid/mp_uuid.h" /* mp_decode_uuid() */
-#include "lib/core/mp_extension_types.h"
+#include "mp_extension_types.h"
+#include "mp_uuid.h" /* mp_decode_uuid() */
 
 #include "cord_buf.h"
 #include <fiber.h>
