+++ test/sql-tap/CMakeLists.txt
@@ -4,6 +4,6 @@ target_link_libraries(gh-5938-wrong-string-length msgpuck)
 build_module(gh-6024-funcs-return-bin gh-6024-funcs-return-bin.c)
 target_link_libraries(gh-6024-funcs-return-bin msgpuck)
 build_module(sql_uuid sql_uuid.c)
-target_link_libraries(sql_uuid msgpuck uuid)
+target_link_libraries(sql_uuid msgpuck core)
 build_module(decimal decimal.c)
 target_link_libraries(decimal msgpuck)
