+++ test/unit/CMakeLists.txt
@@ -53,7 +53,7 @@ target_link_libraries(bitset_index.test bitset)
 add_executable(base64.test base64.c)
 target_link_libraries(base64.test misc unit)
 add_executable(uuid.test uuid.c core_test_utils.c)
-target_link_libraries(uuid.test uuid unit)
+target_link_libraries(uuid.test core unit)
 add_executable(random.test random.c core_test_utils.c)
 target_link_libraries(random.test core unit)
 add_executable(xmalloc.test xmalloc.c core_test_utils.c)
