+++ test/unit/vy_iterators_helper.c
@@ -1,8 +1,8 @@
 #include "vy_iterators_helper.h"
 #include "memory.h"
 #include "fiber.h"
-#include "uuid/tt_uuid.h"
 #include "say.h"
+#include "tt_uuid.h"
 
 struct tt_uuid INSTANCE_UUID;
